form/requirements.txt
Flask
wtforms
requests
pymysql
PyYAML
